Setting the canvas to the videos dimensions.
If you manually set the canvas dimension then there is an issue when
you switch to a camera which has a different aspect ratio (e.g. on
my Nexus 7 both cameras are portrait rather than landscape).

If you don't set the width and height that you are pulling from the
video then only a subsection of the video feed is scanned for QR codes
(see 39e7e0a36f76dbcf85cc37d6d45eb92cb409c059).

This seems to me to be the only way to deal with it although I now
wonder if you need to call prepareCanvas manually at all...